# src/main/java/com/eteosf/teammanager/model/dto/PlayerDto.java
package com.eteosf.teammanager.model.dto;

public class PlayerDto {
    private Long id;
    private String name;
    private Integer age;

    public PlayerDto() {
    }

    public PlayerDto(Long id, String name, Integer age) {
        this.id = id;
        this.name = name;
        this.age = age;
    }

    public Long getId() {
        return id;
    }

    public void setId(Long id) {
        this.id = id;
    }

    public String getName() {
        return name;
    }

    public void setName(String name) {
        this.name = name;
    }

    public Integer getAge() {
        return age;
    }

    public void setAge(Integer age) {
        this.age = age;
    }

    @Override
    public boolean equals(Object o) {
        if (this == o) return true;
        if (o == null || getClass() != o.getClass()) return false;

        PlayerDto playerDto = (PlayerDto) o;

        if (age != playerDto.age) return false;
        return name != null ? name.equals(playerDto.name) : playerDto.name == null;
    }

    @Override
    public int hashCode() {
        int result = name != null ? name.hashCode() : 0;
        result = 31 * result + age;
        return result;
    }

    @Override
    public String toString() {
        return "PlayerDto{" +
                "name='" + name + '\'' +
                ", age=" + age +
                '}';
    }
}